Restructured drilldownModule, removed old subProcessModule
