+++ demo.html
@@ -22,7 +22,8 @@
       additionalModules: [
         bpmnViewer.customModules.MoveCanvasModule,
         bpmnViewer.customModules.ZoomScrollModule,
-        bpmnViewer.customModules.customDrilldownModule,
+        bpmnViewer.customModules.callActivityDrilldownModule,
+        bpmnViewer.customModules.drilldownCenteringModule,
         bpmnViewer.customModules.styleModule
       ],
       bpmnRenderer: {
