+++ index.js
@@ -2,7 +2,8 @@ import Viewer from 'bpmn-js/lib/Viewer';
 
 import MoveCanvasModule from 'diagram-js/lib/navigation/movecanvas';
 import ZoomScrollModule from 'diagram-js/lib/navigation/zoomscroll';
-import customDrilldownModule from './lib/customDrilldown';
+import callActivityDrilldownModule from './lib/customDrilldown/callActivityDrilldown';
+import drilldownCenteringModule from './lib/customDrilldown/drilldownCentering';
 import styleModule from './lib/styleModule';
 
 var bpmnViewer = {
@@ -10,7 +11,8 @@ var bpmnViewer = {
   customModules: {
     MoveCanvasModule,
     ZoomScrollModule,
-    customDrilldownModule,
+    callActivityDrilldownModule,
+    drilldownCenteringModule,
     styleModule
   },
 };
