+++ lib/customDrilldown/callActivityDrilldown/index.js
@@ -1,0 +1,6 @@
+import CallActivityDrilldown from './CallActivityDrilldown';
+
+export default {
+  __init__: ['callActivityDrilldown'],
+  callActivityDrilldown: ['type', CallActivityDrilldown],
+};
