+++ lib/customDrilldown/drilldownCentering/index.js
@@ -1,0 +1,6 @@
+import DrilldownCentering from './DrilldownCentering';
+
+export default {
+  __init__: ['drilldownCentering'],
+  drilldownCentering: ['type', DrilldownCentering],
+};
