--- lib/subProcessModule/index.js
@@ -1,6 +1,0 @@
-import SubProcessModule from './SubProcessModule.js';
-
-export default {
-  __init__: ['subProcessModule'],
-  subProcessModule: ['type', SubProcessModule],
-};
